+++ device/tests/Microsoft.Azure.Devices.Client.Test/Microsoft.Azure.Devices.Client.Test.csproj
@@ -59,6 +59,7 @@
   </Choose>
   <ItemGroup>
     <Compile Include="AmqpTransportHandlerTests.cs" />
+    <Compile Include="AuthenticationWithTokenRefreshTests.cs" />
     <Compile Include="CertificateHelper.cs" />
     <Compile Include="AuthenticationMethod\ModuleAuthenticationTests.cs" />
     <Compile Include="ModuleAuthenticationWithTokenRefreshTests.cs" />
